requirements.txt
# 🎓 课程助手 - 智能问答系统依赖列表
# ========================================
# 基于RAG技术的课程内容智能问答平台
# 项目地址: https://github.com/yourusername/course-assistant-rag
# Python版本要求: >=3.8

# 核心框架
# ========================================
gradio>=4.44.1              # Web界面框架
langchain>=0.3.7             # RAG框架核心
langchain-community>=0.3.5   # LangChain社区组件
langchain-core>=0.3.15       # LangChain核心组件

# 向量数据库和搜索
# ========================================
faiss-cpu>=1.8.0             # 向量相似性搜索（CPU版本）
rank-bm25>=0.2.2             # BM25关键词搜索
sentence-transformers>=3.3.1 # 句子嵌入模型

# 文档处理
# ========================================
pypdf>=5.1.0                 # PDF文档处理
python-docx>=1.1.2           # Word文档处理
python-pptx>=0.6.21          # PowerPoint文档处理
openpyxl>=3.1.2              # Excel文档处理
markdown>=3.5.1              # Markdown文档处理
beautifulsoup4>=4.12.2       # HTML文档处理
lxml>=4.9.3                  # XML文档处理

# 多媒体处理
# ========================================
Pillow>=10.4.0               # 图像处理
pytesseract>=0.3.13          # OCR文字识别
opencv-python>=4.8.1         # 计算机视觉

# 数据处理和科学计算
# ========================================
pandas>=1.3.0                # 数据分析
numpy>=1.21.0                # 数值计算
scikit-learn>=1.3.2          # 机器学习工具

# 网络请求和API
# ========================================
requests>=2.32.3             # HTTP请求
openai>=1.54.4               # OpenAI API客户端
httpx>=0.25.2                # 异步HTTP客户端

# 中文文本处理
# ========================================
jieba>=0.42.1                # 中文分词

# 环境配置和工具
# ========================================
python-dotenv>=1.0.1         # 环境变量管理
pydantic>=2.5.0              # 数据验证
click>=8.1.7                 # 命令行工具
rich>=13.9.4                 # 富文本终端显示
tqdm>=4.65.0                 # 进度条
psutil>=5.9.5                # 系统监控

# 开发和测试工具（可选）
# ========================================
pytest>=7.4.3               # 单元测试框架
black>=23.11.0               # 代码格式化
flake8>=6.1.0                # 代码质量检查
mypy>=1.7.0                  # 静态类型检查

# 部署相关（可选）
# ========================================
gunicorn>=21.2.0             # WSGI HTTP服务器
uvicorn>=0.24.0              # ASGI服务器

# 注意事项:
# 1. 如需GPU加速，请将 faiss-cpu 替换为 faiss-gpu
# 2. 某些包可能需要额外的系统依赖，请参考官方文档
# 3. 建议使用虚拟环境安装依赖，避免版本冲突
